src/third_party/WebKit/LayoutTests/http/tests/htmlimports/import-onload-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
PASS ['bye', 'hello1', 'hello2', 'hello3'] is givenLoadEventTargets.sort()
PASS successfullyParsed is true

TEST COMPLETE
